pyproject.toml
[project]
name = "dp-filtering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "huggingface-hub>=0.36.0",
    "matplotlib>=3.10.8",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "transformers>=4.57.5",
]
